output warning if uploaders files contain key it most likely too long
gpgme only outputs a keyid that currently has 16 hex digits. That means
one either would have to let gpgme check every key id given in uploaders
files (which would then most likely call gpg for every single one), or
the keys given must match what gpgme is returning in key signatures.

The later is the way it is implemented which has the disadvantage that
a longer keyid is not matched. Making this a hard error would cause
problems in case gpgme outputs longer keyids in the future, so instead
just print a suppressable warning.